pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "browser-use>=0.1.24",
    "chainlit>=2.4.400",
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "fastembed>=0.6.1",
    "goat-sdk>=0.1.6",
    "goat-sdk-adapter-langchain>=0.1.1",
    "goat-sdk-plugin-erc20>=0.1.2",
    "goat-sdk-wallet-evm>=0.1.2",
    "goat-sdk-wallet-web3>=0.1.5",
    "google-cloud-aiplatform>=1.88.0",
    "google-cloud-discoveryengine>=0.13.8",
    "influxdb-client>=1.48.0",
    "langchain>=0.3.23",
    "langchain-community>=0.3.21",
    "langchain-core>=0.3.51",
    "langchain-google-genai>=2.1.2",
    "langchain-google-vertexai>=2.0.19",
    "langchain-openai>=0.3.12",
    "langchain-upstage>=0.6.0",
    "langgraph>=0.3.28",
    "playwright>=1.51.0",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "qdrant-client>=1.13.3",
    "requests>=2.32.3",
    "uvicorn[standard]>=0.34.0",
    "vertexai>=1.43.0",
]
